Refactor by divide FirebaseRepo to Messaging and Auth repo
