+++ data/src/main/java/com/foundy/data/di/RepositoryModule.kt
@@ -35,8 +35,14 @@ class RepositoryModule {
 
     @Provides
     @Singleton
-    fun provideFirebaseRepository(): FirebaseRepository {
-        return FirebaseRepositoryImpl()
+    fun provideAuthRepository(): AuthRepository {
+        return AuthRepositoryImpl()
+    }
+
+    @Provides
+    @Singleton
+    fun provideMessagingRepository(): MessagingRepository {
+        return MessagingRepositoryImpl()
     }
 
     @Provides
