+++ domain/src/main/java/com/foundy/domain/repository/AuthRepository.kt
@@ -1,0 +1,8 @@
+package com.foundy.domain.repository
+
+import androidx.lifecycle.LiveData
+
+interface AuthRepository {
+    fun isSignedIn(): Boolean
+    fun signInWith(idToken: String): LiveData<Result<Any>>
+}
