+++ domain/src/main/java/com/foundy/domain/repository/MessagingRepository.kt
@@ -1,10 +1,6 @@
 package com.foundy.domain.repository
 
-import androidx.lifecycle.LiveData
-
-interface FirebaseRepository {
-    fun isSignedIn(): Boolean
-    fun signInWith(idToken: String): LiveData<Result<Any>>
+interface MessagingRepository {
     fun subscribeAllPreviousTopic()
     fun subscribeTo(topic: String, onFailure: (Exception) -> Unit)
     fun unsubscribeFrom(topic: String, onFailure: (Exception) -> Unit)
