+++ domain/src/main/java/com/foundy/domain/usecase/auth/IsSignedInUseCase.kt
@@ -1,0 +1,10 @@
+package com.foundy.domain.usecase.auth
+
+import com.foundy.domain.repository.AuthRepository
+import javax.inject.Inject
+
+class IsSignedInUseCase @Inject constructor(
+    private val repository: AuthRepository
+) {
+    operator fun invoke() = repository.isSignedIn()
+}
