+++ domain/src/main/java/com/foundy/domain/usecase/auth/SignInWithUseCase.kt
@@ -1,10 +1,10 @@
-package com.foundy.domain.usecase.firebase
+package com.foundy.domain.usecase.auth
 
-import com.foundy.domain.repository.FirebaseRepository
+import com.foundy.domain.repository.AuthRepository
 import javax.inject.Inject
 
 class SignInWithUseCase @Inject constructor(
-    private val repository: FirebaseRepository
+    private val repository: AuthRepository
 ) {
     operator fun invoke(idToken: String) = repository.signInWith(idToken)
 }
