--- domain/src/main/java/com/foundy/domain/usecase/firebase/IsSignedInUseCase.kt
@@ -1,10 +1,0 @@
-package com.foundy.domain.usecase.firebase
-
-import com.foundy.domain.repository.FirebaseRepository
-import javax.inject.Inject
-
-class IsSignedInUseCase @Inject constructor(
-    private val repository: FirebaseRepository
-) {
-    operator fun invoke() = repository.isSignedIn()
-}
