+++ presentation/src/main/java/com/foundy/presentation/view/keyword/LoginViewModel.kt
@@ -1,8 +1,8 @@
 package com.foundy.presentation.view.keyword
 
 import androidx.lifecycle.ViewModel
-import com.foundy.domain.usecase.firebase.SignInWithUseCase
-import com.foundy.domain.usecase.firebase.SubscribeAllPreviousTopicUseCase
+import com.foundy.domain.usecase.auth.SignInWithUseCase
+import com.foundy.domain.usecase.messaging.SubscribeAllPreviousTopicUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
